Cleanup: ลบไฟล์เก่าและปรับปรุง module ใหม่
