+++ cmd/api/main.go
@@ -10,10 +10,11 @@ import (
 	"github.com/gofiber/fiber/v2"
 	"github.com/joho/godotenv"
 
+	"go-template/internal/adapters/primary/http/handlers"
 	"go-template/internal/adapters/primary/http/middleware"
-	exampleModule "go-template/internal/modules/example-module"
+	"go-template/internal/modules/example_module"
 	"go-template/pkg/config"
-	"go-template/pkg/platform"
+	postgres "go-template/pkg/platform/postgres"
 )
 
 func main() {
@@ -30,17 +31,42 @@ func main() {
 		log.Fatalf("failed to load configuration: %v", err)
 	}
 
-	// 2. เชื่อมต่อ Database (แบบ DI)
-	db, err := platform.InitPostgres() // ✨ ใช้วิธีที่มีจริง
+	// 2. เชื่อมต่อ Database หลักฐาน (Primary Database - Required)
+	primaryDB, err := postgres.NewConnection(cfg.Database)
 	if err != nil {
-		log.Fatalf("Failed to connect to database: %v", err)
+		log.Fatalf("Failed to connect to primary database: %v", err)
 	}
-	log.Println("✅ Successfully connected to PostgreSQL")
+	log.Println("✅ Primary database connected")
 
-	// 3. ประกอบร่าง Modules (Dependency Injection)
-	exampleRepo := exampleModule.NewExampleRepository(db)
-	exampleService := exampleModule.NewExampleService(exampleRepo)
-	exampleHandler := exampleModule.NewExampleHandler(exampleService)
+	// 3. เชื่อมต่อ Analytics Database (Optional)
+	analyticsDB, err := postgres.InitPostgresWithName("ANALYTICS")
+	if err != nil {
+		log.Printf("⚠️ Analytics database not available: %v", err)
+		analyticsDB = nil // ใช้ nil แทน
+	} else {
+		log.Println("✅ Analytics database connected")
+	}
+
+	// 4. เชื่อมต่อ Logs Database (Optional)
+	logsDB, err := postgres.InitPostgresWithName("LOGS")
+	if err != nil {
+		log.Printf("⚠️ Logs database not available: %v", err)
+		logsDB = nil // ใช้ nil แทน
+	} else {
+		log.Println("✅ Logs database connected")
+	}
+
+	// 5. ประกอบร่าง Modules (Dependency Injection)
+	exampleRepo := example_module.NewExampleRepository(primaryDB)
+	exampleService := example_module.NewExampleService(exampleRepo)
+	exampleHandler := example_module.NewExampleHandler(exampleService)
+
+	// ตัวอย่างการใช้งาน multiple databases
+	_ = analyticsDB // ป้องกัน unused variable
+	_ = logsDB      // ป้องกัน unused variable
+
+	// Health handler
+	healthHandler := handlers.NewHealthHandler()
 
 	// 4. ตั้งค่า Web Server (Fiber)
 	app := fiber.New(fiber.Config{
@@ -50,19 +76,14 @@ func main() {
 	app.Use(middleware.Logger())
 	app.Use(middleware.CORS())
 
-	app.Get("/health", func(c *fiber.Ctx) error {
-		return c.JSON(fiber.Map{"status": "ok"})
-	})
+	// Register health routes
+	healthHandler.RegisterRoutes(app)
 
 	// --- ลงทะเบียน Routes ของแต่ละ Module (แบบ Modular) --- ✨
 	apiV1 := app.Group("/api/v1")
 
 	// Register example module routes
-	examples := apiV1.Group("/examples")
-	examples.Post("", exampleHandler.CreateExample)
-	examples.Get("/:id", exampleHandler.GetExample)
-	examples.Put("/:id", exampleHandler.UpdateExample)
-	examples.Delete("/:id", exampleHandler.DeleteExample)
+	exampleHandler.RegisterRoutes(apiV1)
 
 	// 5. เริ่มและปิดการทำงานของ Server (Start & Graceful Shutdown)
 	go func() {
